Modernization a-ok!
Still have to test how it behaves at maximum stats, but the bulk of the process
is running smoothly.
You have to reload the Ship database, because I had mapped the modernization
attribute on Ships to the wrong class...oops.
Also, Woot!! One mystery solved! karyoku on api_ship are the values of modernized stats